requirements.txt
asgiref==3.8.1
attrs==25.1.0
Django==5.1.6
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.2.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
psycopg2==2.9.10
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
argon2-cffi==23.1.0
python-dotenv==1.0.1
gunicorn==22.0.0
whitenoise==6.0.0
django-phonenumber-field==8.0.0
phonenumberslite==9.0.0
pylint==3.3.4
pylint-django==2.6.1
django-allauth==65.4.1
phonenumberslite==9.0.0
pillow==9.1.0
pytest==8.0.0
django-filter==25.1
pytest-django==4.5.2
coverage==7.7.1
